+++ adapters/Chartboost/ChartboostAdapter/ChartboostAdapter-Bridging-Header.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2026 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,14 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "GADMAdapterChartboost.h"
-
-#import "GADMediationAdapterChartboost.h"
-
-@implementation GADMAdapterChartboost
-
-+ (nonnull Class<GADMediationAdapter>)mainAdapterClass {
-  return [GADMediationAdapterChartboost class];
-}
-
-@end
+#import "GADMAdapterChartboostUtils.h"