+++ adapters/Chartboost/ChartboostAdapter/GADMediationAdapterChartboostBannerAd.m
@@ -24,7 +24,6 @@
 
 #import "ChartboostAdapter-Swift.h"
 #import "GADMAdapterChartboostUtils.h"
-#import "GADMChartboostError.h"
 
 @interface GADMediationAdapterChartboostBannerAd () <CHBBannerDelegate>
 @end
@@ -139,7 +138,7 @@
 
 - (void)didCacheAd:(CHBCacheEvent *)event error:(nullable CHBCacheError *)error {
   if (error) {
-    NSError *loadError = GADMChartboostErrorForCHBCacheError(error);
+    NSError *loadError = [GADMChartboostError errorForCacheError:error];
     NSLog(@"Failed to load banner ad from Chartboost: %@", loadError.localizedDescription);
     _completionHandler(nil, loadError);
     return;
@@ -151,7 +150,7 @@
 
 - (void)didShowAd:(CHBShowEvent *)event error:(nullable CHBShowError *)error {
   if (error) {
-    NSError *showError = GADMChartboostErrorForCHBShowError(error);
+    NSError *showError = [GADMChartboostError errorForShowError:error];
     NSLog(@"Failed to show banner ad from Chartboost: %@", showError.localizedDescription);
 
     [_adEventDelegate didFailToPresentWithError:showError];
@@ -162,7 +161,7 @@
 - (void)didClickAd:(CHBClickEvent *)event error:(CHBClickError *)error {
   [_adEventDelegate reportClick];
   if (error) {
-    NSError *clickError = GADMChartboostErrorForCHBClickError(error);
+    NSError *clickError = [GADMChartboostError errorForClickError:error];
     NSLog(@"An error occurred when clicking the Chartboost banner ad: %@",
           clickError.localizedDescription);
   }